Refactore awarding 3.1 structure
